# docs/api-contract.yaml
openapi: "3.0.3"
info:
  title: Web Calculator API
  version: "1.0.0"
  description: REST API for a web calculator with persistent history backed by MongoDB.
servers:
  - url: http://localhost:3001
    description: Local development
paths:
...
  /api/calculate:
    post:
      summary: Evaluate a math expression and persist the result
      operationId: calculate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expression]
              properties:
                expression:
                  type: string
                  minLength: 1
                  example: "2 + 2"
      responses:
        "201":
          description: Calculation saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calculation"
        "400":
          description: Missing or empty expression
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Expression could not be evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Calculation:
      type: object
      required: [id, expression, result, createdAt]
      properties:
        id:
          type: string
          example: "664a1f2e3b8c9d0e1f2a3b4c"
        expression:
          type: string
          example: "2 + 2"
        result:
          type: number
          example: 4
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Expression is required"